fixup: Get rid of numbers, use integers instead
SymEngine is very strange with this as to 0.0 (from SymEngine::number(0)) != 0 and 0.0 isn't exact zero

+++ src/tests/transform/line_segment_concat.cpp
@@ -28,13 +28,13 @@ TEST_CASE("LineSegmentConcat", "[transform]") {
 
     system.Execute([](System& system) { return nullptr; });
 
-    REQUIRE(system.Algebra().CheckEquation(LineSegmentLength(p2, p3) - SymEngine::number(2)));
+    REQUIRE(system.Algebra().CheckEquation(LineSegmentLength(p2, p3) - SymEngine::integer(2)));
 
     // Let: AD=8
     system.Algebra().AddEquation(LineSegmentLength(p1, p4) - SymEngine::integer(8));
 
-    REQUIRE(system.Algebra().CheckEquation(LineSegmentLength(p3, p4) - SymEngine::number(3)));
-    REQUIRE(system.Algebra().CheckEquation(LineSegmentLength(p2, p4) - SymEngine::number(5)));
+    REQUIRE(system.Algebra().CheckEquation(LineSegmentLength(p3, p4) - SymEngine::integer(3)));
+    REQUIRE(system.Algebra().CheckEquation(LineSegmentLength(p2, p4) - SymEngine::integer(5)));
 }
 
 } // namespace Core
